[TASK] handlers moved to core bl
